Move EIT* to informedtrees folder
