+++ src/ompl/geometric/planners/informedtrees/eitstar/Direction.h
@@ -34,7 +34,8 @@
 
 // Authors: Marlin Strub
 
-#include "ompl/geometric/planners/eitstar/stopwatch/timetable.h"
+#ifndef OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_EITSTAR_DIRECTION_
+#define OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_EITSTAR_DIRECTION_
 
 namespace ompl
 {
@@ -42,48 +43,15 @@ namespace ompl
     {
         namespace eitstar
         {
-            namespace timing
-            {
-                template <>
-                const std::string Timetable<std::chrono::nanoseconds>::units()
-                {
-                    return "nanoseconds";
-                }
-
-                template <>
-                const std::string Timetable<std::chrono::microseconds>::units()
-                {
-                    return "microseconds";
-                }
-
-                template <>
-                const std::string Timetable<std::chrono::milliseconds>::units()
-                {
-                    return "milliseconds";
-                }
-
-                template <>
-                const std::string Timetable<std::chrono::seconds>::units()
-                {
-                    return "seconds";
-                }
-
-                template <>
-                const std::string Timetable<std::chrono::minutes>::units()
-                {
-                    return "minutes";
-                }
-
-                template <>
-                const std::string Timetable<std::chrono::hours>::units()
-                {
-                    return "hours";
-                }
-
-            }  // namespace timing
+            enum class Direction {
+                FORWARD,
+                REVERSE
+            };
 
         }  // namespace aibitstar
 
     }  // namespace geometric
 
 }  // namespace ompl
+
+#endif  // OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_EITSTAR_DIRECTION_
