+++ src/ompl/geometric/planners/informedtrees/eitstar/src/ReverseQueue.cpp
@@ -34,12 +34,12 @@
 
 // Authors: Marlin Strub
 
-#include "ompl/geometric/planners/eitstar/ReverseQueue.h"
+#include "ompl/geometric/planners/informedtrees/eitstar/ReverseQueue.h"
 
 #include <utility>
 
-#include "ompl/geometric/planners/eitstar/Direction.h"
-#include "ompl/geometric/planners/eitstar/State.h"
+#include "ompl/geometric/planners/informedtrees/eitstar/Direction.h"
+#include "ompl/geometric/planners/informedtrees/eitstar/State.h"
 
 namespace ompl
 {
